requirements.txt
bcrypt
cffi
cryptography
future
Jinja2
MarkupSafe
netmiko
ntc-templates
paramiko
pip
pycparser
PyNaCl
pyserial
python-dotenv
PyYAML
scp
setuptools
six
textfsm
